repo-updater: Move GithubSource test to github_test.go
